requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.0.1 